common: remove Cgroups v1 support
Removes all conditionals for IsCgroup2UnifiedMode and cgroupv2.Enabled
and assumes Cgroups v2. Also removes `cgroup2` and `additionalControllers` from
CgroupControl struct and all the conditionals resulting from it.

pkg/cgroups.IsCgroup2UnifiedMode is being retained as it will be used in
the dependent tools as a single check for v1/v2.

Fixes: RUN-3567 (partly)

Signed-off-by: Lokesh Mandvekar <lsm5@redhat.com>