libartifact: strong type where possible
introduction of ArtifactReference and ArtifactStorageReference types for
libartifact.  ArtifactReference is a theoretical reference to an
artifact like for things like pull.  ArtifactStorageReference is for
looking things up in the existing store for things like rm or inspect.
ArtifactStoreReference allows for things like full or partial "id"
lookups.

the pr also accomdates a user ask for appending "latest" as a tag when
no tag is provided.

and finally, here we introduce a bunch of artifact tests to get the ball
rolling.  it should not be considered fully complete but a good start.

Signed-off-by: Brent Baude <bbaude@redhat.com>