Strong type libartifact
Created two new types ArtifactReference and ArtifactStorageReference and
switch the relevant exported functions to use these types.

ArtifactReference is a theoretical reference to an artifact.  It is
essentially a named.  As such, anything a named can be, so can this.
This type is used in functions like pull and add.

ArtifactStorageReference is also a theoretical reference to an artifact
but to an artifact in the local store.  The input to this type can be
similar to an ArtifactReference.  But it can also be a full or partial
digest (sometimes called an ID) of the artifact in the store.  It is
used in functions like inspect and remove.

GetByNameOrDigest was removed we now lookup artifacts using an
unexported function (for now) as an object of the store.

In my previous PR, issues outside the scope of typing were pointed out
and while they are valid, I would like to address these issues in
separate PRs. This PR does not address:
* any locking issues that exist in artifacts
* tests for the store or primary artifact functions

Signed-off-by: Brent Baude <bbaude@redhat.com>