fix(secrets): replace using correct driver, safe replace for new secrets
When replacing an existing secret, the delete operation should ran on the old driver, which is not necessarily the new one.

When replacing a nonexisting secret (Replace=true), ignore the replace flag, treat it as an entirely new secret. (previous behavior: would attempt to delete a secret whose name or id matches an empty string query, which either failed due to ambiguity or successfully deleted an existing secret if exactly 1 secret was present)

Solves issues 1 and 2 of containers/podman#27130

Signed-off-by: Samuel Gunter <sgunter@utexas.edu>