Add OCI distribution-spec registry proxying support
This commit implements support for the OCI distribution-spec registry
proxying feature as defined in: https://github.com/opencontainers/distribution-spec/blob/main/spec.md#registry-proxying
When a registry acts as a proxy for multiple upstream registries, it needs
to know which upstream registry to route requests to. The distribution-spec
defines the "ns" query parameter for this purpose. For example, a request
to a proxy for docker.io images would include "?ns=docker.io".

Signed-off-by: Sohan Kunkerkar <sohank2602@gmail.com>